moving things into redskull-controller
this is wehre the frontend/main API will live going forward.

For now, you can simply build this from here as you used to. Once it has
been updated to reflect the new architecture this will becoem the
controller code which uses it transparently - you'll just need to
migrate your servers to it which will be easy.